Abstract user info details into a separate component.
