+++ online-shop-app/src/components/common/index.js
@@ -1,1 +1,2 @@
 export { default as Divider } from "./Divider";
+export { default as UserInfo } from "./UserInfo";
