+++ online-shop-app/src/components/index.js
@@ -1,4 +1,4 @@
 import * as CustomerFlowComponents from "./customer-flow";
 
 export { CustomerFlowComponents };
-export { Divider } from "./common"
+export { Divider, UserInfo } from "./common"
